third_party/blink/web_tests/platform/generic/fast/css/invalidation/targeted-hover-invalidation-expected.txt

Check that invalidation sets work for :hover changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.internals is defined.
PASS window.eventSender is defined.
PASS getComputedStyle(r1, null).backgroundColor is transparent
PASS getComputedStyle(r2, null).backgroundColor is transparent
PASS r3.length is 2
PASS getComputedStyle(r3[0], null).backgroundColor is transparent
PASS getComputedStyle(r3[1], null).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(r1, null).backgroundColor is green
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(r2, null).backgroundColor is green
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS getComputedStyle(r3[0], null).backgroundColor is green
PASS getComputedStyle(r3[1], null).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
